+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Option added to display greater mechanical detail of the clock.
 * Option added for emulation of mechanical timekeeping (supersedes “Align sun to hand pointer” from version 1.2.0).
 * Option for sun/moon "real position" markers.
+* Option to show addition planet positions (Mercury, Venus, Mars, Jupiter, Saturn).
 * For greater consistency with post-2018 appearance, globe map can be hidden.
 * Collapsed control panel displays currently-set time.
 * Option added for keeping SVG shading at the risk of slower graphics.